Fix: check if table exists before dropping

+++ db/migrate/20130724121010_drop_markets.rb
@@ -1,6 +1,10 @@
 class DropMarkets < ActiveRecord::Migration
   def change
-    drop_table :markets
-    drop_table :markets_startups
+    if ActiveRecord::Base.connection.table_exists? 'markets'
+      drop_table :markets
+    end
+    if ActiveRecord::Base.connection.table_exists? 'markets_startups'
+      drop_table :markets_startups
+    end
   end
 end
